Add slab mode classes
The `AbstractSlabMode` class is used for the data common to both TE and
TM modes, and the visualization of the material properties.

The `TESlabMode` and `TMSlabMode`, instead, hold the mode data and its
plotting methods.

It would be interesting to add also the Poynting vector evaluation and
the overlap integral.

These classes are currently unused.